projects/movie-rater/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
charset-normalizer==3.3.0
fastapi==0.103.2
greenlet==2.0.2
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
